Divide user tests by common concept(ish)
