+++ spec/models/user_part_database_and_cache_spec.rb
@@ -1,38 +1,22 @@
-require 'ostruct'
+#require 'ostruct'
 require_relative '../spec_helper'
-require_relative 'user_shared_examples'
-require_relative '../../services/dataservices-metrics/lib/isolines_usage_metrics'
-require_relative '../../services/dataservices-metrics/lib/observatory_snapshot_usage_metrics'
-require_relative '../../services/dataservices-metrics/lib/observatory_general_usage_metrics'
+#require_relative 'user_shared_examples'
+#require_relative '../../services/dataservices-metrics/lib/isolines_usage_metrics'
+#require_relative '../../services/dataservices-metrics/lib/observatory_snapshot_usage_metrics'
+#require_relative '../../services/dataservices-metrics/lib/observatory_general_usage_metrics'
 require 'factories/organizations_contexts'
-require_relative '../../app/model_factories/layer_factory'
-require_dependency 'cartodb/redis_vizjson_cache'
-require 'helpers/rate_limits_helper'
-require 'helpers/unique_names_helper'
+#require_relative '../../app/model_factories/layer_factory'
+#require_dependency 'cartodb/redis_vizjson_cache'
+#require 'helpers/rate_limits_helper'
+#require 'helpers/unique_names_helper'
 require 'helpers/account_types_helper'
-require 'factories/users_helper'
+#require 'factories/users_helper'
 require 'factories/database_configuration_contexts'
 
-describe 'refactored behaviour' do
-  it_behaves_like 'user models' do
-    def get_twitter_imports_count_by_user_id(user_id)
-      get_user_by_id(user_id).get_twitter_imports_count
-    end
-
-    def get_user_by_id(user_id)
-      ::User.where(id: user_id).first
-    end
-
-    def create_user
-      FactoryGirl.create(:valid_user)
-    end
-  end
-end
-
 describe User do
-  include UniqueNamesHelper
+  #include UniqueNamesHelper
   include AccountTypesHelper
-  include RateLimitsHelper
+  #include RateLimitsHelper
 
   before(:each) do
     CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
@@ -67,62 +51,409 @@ describe User do
     @account_type_org.destroy if @account_type_org
   end
 
-  it "should only allow legal usernames" do
-    illegal_usernames = %w(si$mon 'sergio estella' j@vi sergio£££ simon_tokumine SIMON Simon jose.rilla -rilla rilla-)
-    legal_usernames   = %w(simon javier-de-la-torre sergio-leiva sergio99)
+  it "should set default statement timeout values" do
+    @user.in_database["show statement_timeout"].first[:statement_timeout].should == "5min"
+    @user.in_database(as: :public_user)["show statement_timeout"].first[:statement_timeout].should == "5min"
+  end
 
-    illegal_usernames.each do |name|
-      @user.username = name
-      @user.valid?.should be_false
-      @user.errors[:username].should be_present
+  it "should keep in sync user statement_timeout" do
+    @user.user_timeout = 1000000
+    @user.database_timeout = 300000
+    @user.save
+    @user.in_database["show statement_timeout"].first[:statement_timeout].should == "1000s"
+    @user.in_database(as: :public_user)["show statement_timeout"].first[:statement_timeout].should == "5min"
+  end
+
+  it "should keep in sync database statement_timeout" do
+    @user.user_timeout = 300000
+    @user.database_timeout = 1000000
+    @user.save
+    @user.in_database["show statement_timeout"].first[:statement_timeout].should == "5min"
+    @user.in_database(as: :public_user)["show statement_timeout"].first[:statement_timeout].should == "1000s"
+  end
+
+  it "should has their own database, created when the account is created" do
+    @user.database_name.should == "cartodb_test_user_#{@user.id}_db"
+    @user.database_username.should == "test_cartodb_user_#{@user.id}"
+    @user.in_database.test_connection.should == true
+  end
+
+  it 'creates an importer schema in the user database' do
+    @user.in_database[%Q(SELECT * FROM pg_namespace)]
+      .map { |record| record.fetch(:nspname) }
+      .should include 'cdb_importer'
+  end
+
+  it 'creates a cdb schema in the user database' do
+    pending "I believe cdb schema was never used"
+    @user.in_database[%Q(SELECT * FROM pg_namespace)]
+      .map { |record| record.fetch(:nspname) }
+      .should include 'cdb'
+  end
+
+  it 'allows access to the importer schema by the owner' do
+    @user.in_database.run(%Q{
+      CREATE TABLE cdb_importer.bogus ( bogus varchar(40) )
+    })
+    query = %Q(SELECT * FROM cdb_importer.bogus)
+
+    expect { @user.in_database(as: :public_user)[query].to_a }
+      .to raise_error(Sequel::DatabaseError)
+
+    @user.in_database[query].to_a
+  end
+
+  it 'allows access to the cdb schema by the owner' do
+    pending "I believe cdb schema was never used"
+    @user.in_database.run(%Q{
+      CREATE TABLE cdb.bogus ( bogus varchar(40) )
+    })
+    query = %Q(SELECT * FROM cdb.bogus)
+
+    expect { @user.in_database(as: :public_user)[query].to_a }
+      .to raise_error(Sequel::DatabaseError)
+
+    @user.in_database[query].to_a
+  end
+
+  it "should create a dabase user that only can read it's own database" do
+
+    connection = ::Sequel.connect(
+      ::SequelRails.configuration.environment_for(Rails.env).merge(
+        'database' => @user.database_name, :logger => ::Rails.logger,
+        'username' => @user.database_username, 'password' => @user.database_password
+      )
+    )
+    connection.test_connection.should == true
+    connection.disconnect
+
+    connection = nil
+    connection = ::Sequel.connect(
+      ::SequelRails.configuration.environment_for(Rails.env).merge(
+        'database' => @user2.database_name, :logger => ::Rails.logger,
+        'username' => @user.database_username, 'password' => @user.database_password
+      )
+    )
+    begin
+      connection.test_connection
+      true.should_not be_true
+    rescue
+      true.should be_true
+    ensure
+      connection.disconnect
     end
 
-    legal_usernames.each do |name|
-      @user.username = name
-      @user.valid?.should be_true
-      @user.errors[:username].should be_blank
+    connection = ::Sequel.connect(
+      ::SequelRails.configuration.environment_for(Rails.env).merge(
+        'database' => @user2.database_name, :logger => ::Rails.logger,
+        'username' => @user2.database_username, 'password' => @user2.database_password
+      )
+    )
+    connection.test_connection.should == true
+    connection.disconnect
+
+    connection = ::Sequel.connect(
+      ::SequelRails.configuration.environment_for(Rails.env).merge(
+        'database' => @user.database_name, :logger => ::Rails.logger,
+        'username' => @user2.database_username, 'password' => @user2.database_password
+      )
+    )
+    begin
+      connection.test_connection
+      true.should_not be_true
+    rescue
+      true.should be_true
+    ensure
+      connection.disconnect
     end
   end
 
-  it "should not allow a username in use by an organization" do
-    org = create_org('testusername', 10.megabytes, 1)
-    @user.username = org.name
-    @user.valid?.should be_false
-    @user.username = 'wadus'
-    @user.valid?.should be_true
+  it "should run valid queries against their database" do
+    # initial select tests
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='Polynoidae' limit 10")
+    query_result[:time].should_not be_blank
+    query_result[:time].to_s.match(/^\d+\.\d+$/).should be_true
+    query_result[:total_rows].should == 2
+    query_result[:rows].first.keys.sort.should == [:cartodb_id, :the_geom, :the_geom_webmercator, :id, :name_of_species, :kingdom, :family, :lat, :lon, :views].sort
+    query_result[:rows][0][:name_of_species].should == "Barrukia cristata"
+    query_result[:rows][1][:name_of_species].should == "Eulagisca gigantea"
+
+    # update and reselect
+    query_result = @user.db_service.run_pg_query("update import_csv_1 set family='polynoidae' where family='Polynoidae'")
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='Polynoidae' limit 10")
+    query_result[:total_rows].should == 0
+
+    # check counts
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='polynoidae' limit 10")
+    query_result[:total_rows].should == 2
+
+    # test a product
+    query_result = @user.db_service.run_pg_query("select import_csv_1.family as fam, twitters.login as login from import_csv_1, twitters where family='polynoidae' limit 10")
+    query_result[:total_rows].should == 10
+    query_result[:rows].first.keys.should == [:fam, :login]
+    query_result[:rows][0].should == { :fam=>"polynoidae", :login=>"vzlaturistica " }
+
+    # test counts
+    query_result = @user.db_service.run_pg_query("select count(*) from import_csv_1 where family='polynoidae' ")
+    query_result[:time].should_not be_blank
+    query_result[:time].to_s.match(/^\d+\.\d+$/).should be_true
+    query_result[:total_rows].should == 1
+    query_result[:rows].first.keys.should ==  [:count]
+    query_result[:rows][0].should == {:count => 2}
   end
 
+  it "should raise errors when running invalid queries against their database" do
+    lambda {
+      @user.db_service.run_pg_query("selectttt * from import_csv_1 where family='Polynoidae' limit 10")
+    }.should raise_error(CartoDB::ErrorRunningQuery)
+  end
 
-  describe '#destroy_restrictions' do
-    it 'Checks some scenarios upon user destruction regarding organizations' do
-      u1 = create_user(email: 'u1@example.com', username: 'u1', password: 'admin123')
-      u2 = create_user(email: 'u2@example.com', username: 'u2', password: 'admin123')
+  it "should run valid queries against their database in pg mode" do
+    reload_user_data(@user) && @user.reload
 
-      org = create_org('cartodb', 1234567890, 5)
+    # initial select tests
+    # tests results and modified flags
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='Polynoidae' limit 10")
+    query_result[:time].should_not be_blank
+    query_result[:time].to_s.match(/^\d+\.\d+$/).should be_true
+    query_result[:total_rows].should == 2
+    query_result[:rows].first.keys.sort.should == [:cartodb_id, :the_geom, :the_geom_webmercator, :id, :name_of_species, :kingdom, :family, :lat, :lon, :views].sort
+    query_result[:rows][0][:name_of_species].should == "Barrukia cristata"
+    query_result[:rows][1][:name_of_species].should == "Eulagisca gigantea"
+    query_result[:results].should  == true
+    query_result[:modified].should == false
 
-      u1.organization = org
-      u1.save
-      u1.reload
-      u1.organization.nil?.should eq false
-      org = u1.organization
-      org.owner_id = u1.id
-      org.save
-      u1.reload
-      u1.organization.owner.id.should eq u1.id
+    # update and reselect
+    query_result = @user.db_service.run_pg_query("update import_csv_1 set family='polynoidae' where family='Polynoidae'")
+    query_result[:modified].should   == true
+    query_result[:results].should    == false
 
-      u2.organization = org
-      u2.save
-      u2.reload
-      u2.organization.nil?.should eq false
-      u2.reload
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='Polynoidae' limit 10")
+    query_result[:total_rows].should == 0
+    query_result[:modified].should   == false
+    query_result[:results].should    == true
 
-      # Cannot remove as more users depend on the org
-      expect {
-        u1.destroy
-      }.to raise_exception CartoDB::BaseCartoDBError
+    # # check counts
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='polynoidae' limit 10")
+    query_result[:total_rows].should == 2
+    query_result[:results].should    == true
 
-      org.destroy
+    # test a product
+    query_result = @user.db_service.run_pg_query("select import_csv_1.family as fam, twitters.login as login from import_csv_1, twitters where family='polynoidae' limit 10")
+    query_result[:total_rows].should == 10
+    query_result[:rows].first.keys.should == [:fam, :login]
+    query_result[:rows][0].should == { :fam=>"polynoidae", :login=>"vzlaturistica " }
+
+    # test counts
+    query_result = @user.db_service.run_pg_query("select count(*) from import_csv_1 where family='polynoidae' ")
+    query_result[:time].should_not be_blank
+    query_result[:time].to_s.match(/^\d+\.\d+$/).should be_true
+    query_result[:total_rows].should == 1
+    query_result[:rows].first.keys.should ==  [:count]
+    query_result[:rows][0].should == {:count => 2}
+  end
+
+  it "should raise errors when running invalid queries against their database in pg mode" do
+    lambda {
+      @user.db_service.run_pg_query("selectttt * from import_csv_1 where family='Polynoidae' limit 10")
+    }.should raise_error(CartoDB::ErrorRunningQuery)
+  end
+
+  it "should raise errors when invalid table name used in pg mode" do
+    lambda {
+      @user.db_service.run_pg_query("select * from this_table_is_not_here where family='Polynoidae' limit 10")
+    }.should raise_error(CartoDB::TableNotExists)
+  end
+
+  it "should raise errors when invalid column used in pg mode" do
+    lambda {
+      @user.db_service.run_pg_query("select not_a_col from import_csv_1 where family='Polynoidae' limit 10")
+    }.should raise_error(CartoDB::ColumnNotExists)
+  end
+
+  it "should invalidate all their vizjsons when their account type changes" do
+    @account_type = create_account_type_fg('WADUS')
+    @user.account_type = 'WADUS'
+    CartoDB::Varnish.any_instance.expects(:purge)
+      .with("#{@user.database_name}.*:vizjson").times(1).returns(true)
+    @user.save
+  end
+
+  it "should return the result from the last select query if multiple selects" do
+    reload_user_data(@user) && @user.reload
+
+    query_result = @user.db_service.run_pg_query("select * from import_csv_1 where family='Polynoidae' limit 1; select * from import_csv_1 where family='Polynoidae' limit 10")
+    query_result[:time].should_not be_blank
+    query_result[:time].to_s.match(/^\d+\.\d+$/).should be_true
+    query_result[:total_rows].should == 2
+    query_result[:rows][0][:name_of_species].should == "Barrukia cristata"
+    query_result[:rows][1][:name_of_species].should == "Eulagisca gigantea"
+  end
+
+  it "should allow multiple queries in the format: insert_query; select_query" do
+    query_result = @user.db_service.run_pg_query("insert into import_csv_1 (name_of_species,family) values ('cristata barrukia','Polynoidae'); select * from import_csv_1 where family='Polynoidae' ORDER BY name_of_species ASC limit 10")
+    query_result[:total_rows].should == 3
+    query_result[:rows].map { |i| i[:name_of_species] }.should =~ ["Barrukia cristata", "Eulagisca gigantea", "cristata barrukia"]
+  end
+
+  it "should fail with error if table doesn't exist" do
+    reload_user_data(@user) && @user.reload
+    lambda {
+      @user.db_service.run_pg_query("select * from wadus")
+    }.should raise_error(CartoDB::TableNotExists)
+  end
+
+  it "should have a method that generates users redis users_metadata key" do
+    @user.key.should == "rails:users:#{@user.username}"
+  end
+
+  it "replicates some user metadata in redis after saving" do
+    @user.stubs(:database_name).returns('wadus')
+    @user.save
+    $users_metadata.HGET(@user.key, 'id').should == @user.id.to_s
+    $users_metadata.HGET(@user.key, 'database_name').should == 'wadus'
+    $users_metadata.HGET(@user.key, 'database_password').should == @user.database_password
+    $users_metadata.HGET(@user.key, 'database_host').should == @user.database_host
+    $users_metadata.HGET(@user.key, 'map_key').should == @user.api_key
+  end
+
+  it "should store its metadata automatically after creation" do
+    user = FactoryGirl.create :user
+    $users_metadata.HGET(user.key, 'id').should == user.id.to_s
+    $users_metadata.HGET(user.key, 'database_name').should == user.database_name
+    $users_metadata.HGET(user.key, 'database_password').should == user.database_password
+    $users_metadata.HGET(user.key, 'database_host').should == user.database_host
+    $users_metadata.HGET(user.key, 'map_key').should == user.api_key
+    user.destroy
+  end
+
+  it "should have a method that generates users redis limits metadata key" do
+    @user.timeout_key.should == "limits:timeout:#{@user.username}"
+  end
+
+  it "replicates db timeout limits in redis after saving and applies them to db" do
+    @user.user_timeout = 200007
+    @user.database_timeout = 100007
+    @user.save
+    $users_metadata.HGET(@user.timeout_key, 'db').should == '200007'
+    $users_metadata.HGET(@user.timeout_key, 'db_public').should == '100007'
+    @user.in_database do |db|
+      db[%{SHOW statement_timeout}].first.should eq({ statement_timeout: '200007ms' })
     end
+    @user.in_database(as: :public_user) do |db|
+      db[%{SHOW statement_timeout}].first.should eq({ statement_timeout: '100007ms' })
+    end
+  end
+
+  it "replicates render timeout limits in redis after saving" do
+    @user.user_render_timeout = 200001
+    @user.database_render_timeout = 100001
+    @user.save
+    $users_metadata.HGET(@user.timeout_key, 'render').should == '200001'
+    $users_metadata.HGET(@user.timeout_key, 'render_public').should == '100001'
+  end
+
+  it "should store db timeout limits in redis after creation" do
+    user = FactoryGirl.create :user, user_timeout: 200002, database_timeout: 100002
+    user.user_timeout.should == 200002
+    user.database_timeout.should == 100002
+    $users_metadata.HGET(user.timeout_key, 'db').should == '200002'
+    $users_metadata.HGET(user.timeout_key, 'db_public').should == '100002'
+    user.in_database do |db|
+      db[%{SHOW statement_timeout}].first.should eq({ statement_timeout: '200002ms' })
+    end
+    user.in_database(as: :public_user) do |db|
+      db[%{SHOW statement_timeout}].first.should eq({ statement_timeout: '100002ms' })
+    end
+    user.destroy
+  end
+
+  it "should store render timeout limits in redis after creation" do
+    user = FactoryGirl.create :user, user_render_timeout: 200003, database_render_timeout: 100003
+    user.reload
+    user.user_render_timeout.should == 200003
+    user.database_render_timeout.should == 100003
+    $users_metadata.HGET(user.timeout_key, 'render').should == '200003'
+    $users_metadata.HGET(user.timeout_key, 'render_public').should == '100003'
+    user.destroy
+  end
+
+  it "should have valid non-zero db timeout limits by default" do
+    user = FactoryGirl.create :user
+    user.user_timeout.should > 0
+    user.database_timeout.should > 0
+    $users_metadata.HGET(user.timeout_key, 'db').should == user.user_timeout.to_s
+    $users_metadata.HGET(user.timeout_key, 'db_public').should == user.database_timeout.to_s
+    user.in_database do |db|
+      result = db[%{SELECT setting FROM pg_settings WHERE name = 'statement_timeout'}]
+      result.first.should eq(setting: user.user_timeout.to_s)
+    end
+    user.in_database(as: :public_user) do |db|
+      result = db[%{SELECT setting FROM pg_settings WHERE name = 'statement_timeout'}]
+      result.first.should eq(setting: user.database_timeout.to_s)
+    end
+    user.destroy
+  end
+
+  it "should have zero render timeout limits by default" do
+    user = FactoryGirl.create :user
+    user.user_render_timeout.should eq 0
+    user.database_render_timeout.should eq 0
+    $users_metadata.HGET(user.timeout_key, 'render').should eq '0'
+    $users_metadata.HGET(user.timeout_key, 'render_public').should eq '0'
+    user.destroy
+  end
+
+  it "should remove its metadata from redis after deletion" do
+    doomed_user = create_user :email => 'doomed@example.com', :username => 'doomed', :password => 'doomed123'
+    $users_metadata.HGET(doomed_user.key, 'id').should == doomed_user.id.to_s
+    $users_metadata.HGET(doomed_user.timeout_key, 'db').should_not be_nil
+    $users_metadata.HGET(doomed_user.timeout_key, 'db_public').should_not be_nil
+    key = doomed_user.key
+    timeout_key = doomed_user.timeout_key
+    doomed_user.destroy
+    $users_metadata.HGET(key, 'id').should be_nil
+    $users_metadata.HGET(timeout_key, 'db').should be_nil
+    $users_metadata.HGET(timeout_key, 'db_public').should be_nil
+    $users_metadata.HGET(timeout_key, 'render').should be_nil
+    $users_metadata.HGET(timeout_key, 'render_public').should be_nil
+  end
+
+  it "should remove its database and database user after deletion" do
+    doomed_user = create_user :email => 'doomed1@example.com', :username => 'doomed1', :password => 'doomed123'
+    create_table :user_id => doomed_user.id, :name => 'My first table', :privacy => UserTable::PRIVACY_PUBLIC
+    doomed_user.reload
+    SequelRails.connection["select count(*) from pg_catalog.pg_database where datname = '#{doomed_user.database_name}'"]
+      .first[:count].should == 1
+    SequelRails.connection["select count(*) from pg_catalog.pg_user where usename = '#{doomed_user.database_username}'"]
+      .first[:count].should == 1
+
+    doomed_user.destroy
+
+    SequelRails.connection["select count(*) from pg_catalog.pg_database where datname = '#{doomed_user.database_name}'"]
+      .first[:count].should == 0
+    SequelRails.connection["select count(*) from pg_catalog.pg_user where usename = '#{doomed_user.database_username}'"]
+      .first[:count].should == 0
+  end
+
+  it "should invalidate its Varnish cache after deletion" do
+    doomed_user = create_user :email => 'doomed2@example.com', :username => 'doomed2', :password => 'doomed123'
+    CartoDB::Varnish.any_instance.expects(:purge).with("#{doomed_user.database_name}.*").at_least(2).returns(true)
+
+    doomed_user.destroy
+  end
+
+  it "should invalidate all their vizjsons when their disqus_shortname changes" do
+    @user.disqus_shortname = 'WADUS'
+    CartoDB::Varnish.any_instance.expects(:purge)
+      .with("#{@user.database_name}.*:vizjson").times(1).returns(true)
+    @user.save
+  end
+
+  it "should not invalidate anything when their quota_in_bytes changes" do
+    @user.quota_in_bytes = @user.quota_in_bytes + 1.megabytes
+    CartoDB::Varnish.any_instance.expects(:purge).times(0)
+    @user.save
   end
 
   describe '#cartodb_postgresql_extension_versioning' do
@@ -181,8 +512,8 @@ describe User do
       stub_and_check_version_pre_mu('0.2.1 0.2.0-8-g7840e7c', true)
 
       after_mu_old_known_versions = [
-          '0.3.6 0.3.5-8-g7840e7c',
-          '0.4.0 0.3.6-8-g7840e7c'
+        '0.3.6 0.3.5-8-g7840e7c',
+        '0.4.0 0.3.6-8-g7840e7c'
       ]
       after_mu_old_known_versions.each { |version|
         stub_and_check_version_pre_mu(version, false)
@@ -192,256 +523,6 @@ describe User do
     def stub_and_check_version_pre_mu(version, is_pre_mu)
       @user.db_service.stubs(:cartodb_extension_version).returns(version)
       @user.db_service.cartodb_extension_version_pre_mu?.should eq is_pre_mu
-    end
-
-  end
-
-  # INFO: since user can be also created in Central, and it can fail, we need to request notification explicitly. See #3022 for more info
-  it "can notify a new user creation" do
-    ::Resque.stubs(:enqueue).returns(nil)
-    @account_type_org = create_account_type_fg('ORGANIZATION USER')
-    organization = create_organization_with_owner(quota_in_bytes: 1000.megabytes)
-    user1 = new_user(username: 'test',
-                     email: "client@example.com",
-                     organization: organization,
-                     organization_id: organization.id,
-                     quota_in_bytes: 20.megabytes,
-                     account_type: 'ORGANIZATION USER')
-    user1.id = UUIDTools::UUID.timestamp_create.to_s
-
-    ::Resque.expects(:enqueue).with(::Resque::UserJobs::Mail::NewOrganizationUser, user1.id).once
-
-    user1.save
-    # INFO: if user must be synched with a remote server it should happen before notifying
-    user1.notify_new_organization_user
-
-    organization.destroy
-  end
-
-  describe "#change_password" do
-    before(:all) do
-      @new_valid_password = '000123456'
-      @user3 = create_user(password: @user_password)
-    end
-
-    after(:all) do
-      @user3.destroy
-    end
-
-    it "updates crypted_password" do
-      initial_crypted_password = @user3.crypted_password
-      @user3.change_password(@user_password, @new_valid_password, @new_valid_password)
-      @user3.valid?.should eq true
-      @user3.save(raise_on_failure: true)
-      @user3.crypted_password.should_not eql initial_crypted_password
-
-      @user3.change_password(@new_valid_password, @user_password, @user_password)
-      @user3.save(raise_on_failure: true)
-    end
-
-    it "checks old password" do
-      @user3.change_password('aaabbb', @new_valid_password, @new_valid_password)
-      @user3.valid?.should eq false
-      @user3.errors.fetch(:old_password).nil?.should eq false
-      expect {
-        @user3.save(raise_on_failure: true)
-      }.to raise_exception(Sequel::ValidationFailed, /old_password Old password not valid/)
-    end
-
-    it "checks password confirmation" do
-      @user3.change_password(@user_password, 'aaabbb', 'bbbaaa')
-      @user3.valid?.should eq false
-      @user3.errors.fetch(:new_password).nil?.should eq false
-      expect {
-        @user3.save(raise_on_failure: true)
-      }.to raise_exception(Sequel::ValidationFailed, "new_password doesn't match confirmation")
-    end
-
-    it "can throw several errors" do
-      @user3.change_password('aaaaaa', 'aaabbb', 'bbbaaa')
-      @user3.valid?.should eq false
-      @user3.errors.fetch(:old_password).nil?.should eq false
-      @user3.errors.fetch(:new_password).nil?.should eq false
-      expected_errors = "old_password Old password not valid, new_password doesn't match confirmation"
-      expect {
-        @user3.save(raise_on_failure: true)
-      }.to raise_exception(Sequel::ValidationFailed, expected_errors)
-    end
-
-    it "checks minimal length" do
-      @user3.change_password(@user_password, 'tiny', 'tiny')
-      @user3.valid?.should eq false
-      @user3.errors.fetch(:new_password).nil?.should eq false
-      expect {
-        @user3.save(raise_on_failure: true)
-      }.to raise_exception(Sequel::ValidationFailed, "new_password must be at least 6 characters long")
-    end
-
-    it "checks maximal length" do
-      long_password = 'long' * 20
-      @user3.change_password(@user_password, long_password, long_password)
-      @user3.valid?.should eq false
-      @user3.errors.fetch(:new_password).nil?.should eq false
-      expect {
-        @user3.save(raise_on_failure: true)
-      }.to raise_exception(Sequel::ValidationFailed, "new_password must be at most 64 characters long")
-    end
-
-    it "checks that the new password is not nil" do
-      @user3.change_password(@user_password, nil, nil)
-      @user3.valid?.should eq false
-      @user3.errors.fetch(:new_password).nil?.should eq false
-      expect {
-        @user3.save(raise_on_failure: true)
-      }.to raise_exception(Sequel::ValidationFailed, "new_password can't be blank")
-    end
-  end
-
-  describe "when user is signed up with google sign-in and don't have any password yet" do
-    before(:each) do
-      @user.google_sign_in = true
-      @user.last_password_change_date = nil
-      @user.save
-
-      @user.needs_password_confirmation?.should == false
-
-      new_valid_password = '000123456'
-      @user.change_password("doesn't matter in this case", new_valid_password, new_valid_password)
-
-      @user.needs_password_confirmation?.should == true
-    end
-
-    it 'should allow updating password w/o a current password' do
-      @user.valid?.should eq true
-      @user.save
-    end
-
-    it 'should have updated last password change date' do
-      @user.last_password_change_date.should_not eq nil
-      @user.save
-    end
-  end
-
-  describe "#purge_redis_vizjson_cache" do
-    it "shall iterate on the user's visualizations and purge their redis cache" do
-      # Create a few tables with their default vizs
-      (1..3).each do |i|
-        t = Table.new
-        t.user_id = @user.id
-        t.save
-      end
-
-      collection = CartoDB::Visualization::Collection.new.fetch({user_id: @user.id})
-      redis_spy = RedisDoubles::RedisSpy.new
-      redis_vizjson_cache = CartoDB::Visualization::RedisVizjsonCache.new()
-      redis_embed_cache = EmbedRedisCache.new()
-      CartoDB::Visualization::RedisVizjsonCache.any_instance.stubs(:redis).returns(redis_spy)
-      EmbedRedisCache.any_instance.stubs(:redis).returns(redis_spy)
-
-
-      redis_vizjson_keys = collection.map { |v|
-        [
-          redis_vizjson_cache.key(v.id, false), redis_vizjson_cache.key(v.id, true),
-          redis_vizjson_cache.key(v.id, false, 3), redis_vizjson_cache.key(v.id, true, 3),
-          redis_vizjson_cache.key(v.id, false, '3n'), redis_vizjson_cache.key(v.id, true, '3n'),
-          redis_vizjson_cache.key(v.id, false, '3a'), redis_vizjson_cache.key(v.id, true, '3a'),
-        ]
-      }.flatten
-      redis_vizjson_keys.should_not be_empty
-
-      redis_embed_keys = collection.map { |v|
-        [redis_embed_cache.key(v.id, false), redis_embed_cache.key(v.id, true)]
-      }.flatten
-      redis_embed_keys.should_not be_empty
-
-      @user.purge_redis_vizjson_cache
-
-      redis_spy.deleted.should include(*redis_vizjson_keys)
-      redis_spy.deleted.should include(*redis_embed_keys)
-      redis_spy.deleted.count.should eq redis_vizjson_keys.count + redis_embed_keys.count
-      redis_spy.invokes(:del).count.should eq 2
-      redis_spy.invokes(:del).map(&:sort).should include(redis_vizjson_keys.sort)
-      redis_spy.invokes(:del).map(&:sort).should include(redis_embed_keys.sort)
-    end
-
-    it "shall not fail if the user does not have visualizations" do
-      user = create_user
-      collection = CartoDB::Visualization::Collection.new.fetch({user_id: user.id})
-      # 'http' keys
-      redis_keys = collection.map(&:redis_vizjson_key)
-      redis_keys.should be_empty
-      # 'https' keys
-      redis_keys = collection.map { |item| item.redis_vizjson_key(true) }
-      redis_keys.should be_empty
-
-      CartoDB::Visualization::Member.expects(:redis_cache).never
-
-      user.purge_redis_vizjson_cache
-
-      user.destroy
-    end
-  end
-
-  describe "#regressions" do
-    it "Tests geocodings and data import FK not breaking user destruction" do
-      user = create_user
-      user_id = user.id
-
-      data_import_id = '11111111-1111-1111-1111-111111111111'
-
-      SequelRails.connection.run(%Q{
-        INSERT INTO data_imports("data_source","data_type","table_name","state","success","logger","updated_at",
-          "created_at","tables_created_count",
-          "table_names","append","id","table_id","user_id",
-          "service_name","service_item_id","stats","type_guessing","quoted_fields_guessing","content_guessing","server","host",
-          "resque_ppid","upload_host","create_visualization","user_defined_limits")
-          VALUES('test','url','test','complete','t','11111111-1111-1111-1111-111111111112',
-            '2015-03-17 00:00:00.94006+00','2015-03-17 00:00:00.810581+00','1',
-            'test','f','#{data_import_id}','11111111-1111-1111-1111-111111111113',
-            '#{user_id}','public_url', 'test',
-            '[{"type":".csv","size":5015}]','t','f','t','test','0.0.0.0','13204','test','f','{"twitter_credits_limit":0}');
-        })
-
-      SequelRails.connection.run(%Q{
-        INSERT INTO geocodings("table_name","processed_rows","created_at","updated_at","formatter","state",
-          "id","user_id",
-          "cache_hits","kind","geometry_type","processable_rows","real_rows","used_credits",
-          "data_import_id"
-          ) VALUES('importer_123456','197','2015-03-17 00:00:00.279934+00','2015-03-17 00:00:00.536383+00','field_1','finished',
-            '11111111-1111-1111-1111-111111111114','#{user_id}','0','admin0','polygon','195','0','0',
-            '#{data_import_id}');
-        })
-
-      user.destroy
-
-      ::User.find(id:user_id).should eq nil
-
-    end
-  end
-
-  describe '#needs_password_confirmation?' do
-    it 'is true for a normal user' do
-      user = FactoryGirl.build(:carto_user, :google_sign_in => nil)
-      user.needs_password_confirmation?.should == true
-
-      user = FactoryGirl.build(:user, :google_sign_in => false)
-      user.needs_password_confirmation?.should == true
-    end
-
-    it 'is false for users that signed in with Google' do
-      user = FactoryGirl.build(:user, :google_sign_in => true)
-      user.needs_password_confirmation?.should == false
-    end
-
-    it 'is true for users that signed in with Google but changed the password' do
-      user = FactoryGirl.build(:user, :google_sign_in => true, :last_password_change_date => Time.now)
-      user.needs_password_confirmation?.should == true
-    end
-
-    it 'is false for users that were created with http authentication' do
-      user = FactoryGirl.build(:valid_user, last_password_change_date: nil)
-      Carto::UserCreation.stubs(:http_authentication).returns(stub(find_by_user_id: FactoryGirl.build(:user_creation)))
-      user.needs_password_confirmation?.should == false
     end
   end
 
@@ -861,6 +942,62 @@ describe User do
     end
   end
 
+  describe "Write locking" do
+    it "detects locking properly" do
+      @user.db_service.writes_enabled?.should eq true
+      @user.db_service.disable_writes
+      @user.db_service.terminate_database_connections
+      @user.db_service.writes_enabled?.should eq false
+      @user.db_service.enable_writes
+      @user.db_service.terminate_database_connections
+      @user.db_service.writes_enabled?.should eq true
+    end
+
+    it "enables and disables writes in user database" do
+      @user.db_service.run_pg_query("create table foo_1(a int);")
+      @user.db_service.disable_writes
+      @user.db_service.terminate_database_connections
+      lambda {
+        @user.db_service.run_pg_query("create table foo_2(a int);")
+      }.should raise_error(CartoDB::ErrorRunningQuery)
+      @user.db_service.enable_writes
+      @user.db_service.terminate_database_connections
+      @user.db_service.run_pg_query("create table foo_3(a int);")
+    end
+  end
+
+  describe 'ghost tables event trigger' do
+    it 'creates the cdb_ddl_execution table with the user' do
+      table_name = @user.in_database(as: :superuser)
+                     .fetch("SELECT to_regclass('cartodb.cdb_ddl_execution');")
+                     .first[:to_regclass]
+      table_name.should eql 'cdb_ddl_execution'
+    end
+
+    it 'removes the cdb_ddl_execution table when calling drop_ghost_tables_event_trigger' do
+      @user.db_service.drop_ghost_tables_event_trigger
+
+      table_name = @user.in_database(as: :superuser)
+                     .fetch("SELECT to_regclass('cartodb.cdb_ddl_execution');")
+                     .first[:to_regclass]
+      table_name.should be_nil
+
+      @user.db_service.create_ghost_tables_event_trigger
+    end
+
+    it 'saves the TIS config from app_config.yml to cdb_conf' do
+      CartoDB::UserModule::DBService.any_instance.unstub(:configure_ghost_table_event_trigger)
+      user = create_user
+
+      cdb_conf = user.in_database(as: :superuser)
+                   .fetch("SELECT cartodb.CDB_Conf_GetConf('invalidation_service')").first[:cdb_conf_getconf]
+      cdb_conf.should be
+
+      user.destroy
+      CartoDB::UserModule::DBService.any_instance.stubs(:configure_ghost_table_event_trigger).returns(true)
+    end
+  end
+
   protected
 
   def create_org(org_name, org_quota, org_seats)
@@ -872,11 +1009,4 @@ describe User do
     organization
   end
 
-  def tables_including_shared(user)
-    Carto::VisualizationQueryBuilder
-      .new
-      .with_owned_by_or_shared_with_user_id(user.id)
-      .with_type(Carto::Visualization::TYPE_CANONICAL)
-      .build.map(&:table)
-  end
 end
